clean up the formal charges converted from dataframe to change NaN to None

+++ mdtraj/core/topology.py
@@ -649,7 +649,7 @@ class Topology:
             if fc is not None:
                 try:
                     if pd.notna(fc):
-                        fc = int(fc)
+                        fc = float(fc)
                     else:
                         fc = None
                 except Exception:
